op-e2e: Add e2e test for op-program
